Use os.UserHomeDir() (darwin) and drop dependency on go-homedir
This change applies to code compiled only for darwin, but it also allows
to remove a dependency from go.mod which will ease maintenance of
downstream projects for any platforms.

Use os.UserHomeDir() which is available since go1.12 (already the
minimum go version for this project). This allows to drop the dependency
on github.com/mitchellh/go-homedir which was needed only for Darwnin.

go-homedir still has the following features:
* tries very hard to get the location of the home directory when $HOME
  is not set
* avoids use of os/user which dependended on cgo until go1.20

I don't think we really need to go beyond $HOME, but if we really wanted
it we could now use os/user.Current() as since go1.20 it uses the libc on
Darwin and no longer needs cgo on any platform.